rename patch files
