+++ langwatch/src/server/elasticsearch/patchOpensearchCompatibility.ts
