+++ langwatch/src/server/elasticsearch/patchQuickwitCompatibility.ts
